requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
boto3==1.34.110
botocore==1.34.110
Brotli==1.1.0
bs4==0.0.2
build==1.0.3
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpathlib==0.16.0
cohere==5.5.0
coloredlogs==15.0.1
confection==0.1.4
crewai==0.30.11
crewai-tools==0.2.6
cryptography==42.0.7
cymem==2.0.8
dataclasses-json==0.6.3
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
docstring-parser==0.15
docx2txt==0.8
easyocr==1.7.1
embedchain==0.1.102
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.108.0
fastavro==1.9.4
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.19.0
google-auth==2.26.1
google-cloud-aiplatform==1.52.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gptcache==0.1.43
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
httpx-sse==0.4.0
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
imageio==2.34.1
importlib-metadata==6.11.0
importlib-resources==6.1.1
in-n-out==0.2.1
iniconfig==2.0.0
instructor==0.5.2
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
lancedb==0.5.7
langchain==0.1.16
langchain-chroma==0.1.0
langchain-cohere==0.1.5
langchain-community==0.0.32
langchain-core==0.1.52
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langcodes==3.4.0
langsmith==0.1.47
language_data==1.2.0
lazy_loader==0.4
llama-index==0.10.36
llama-index-agent-openai==0.2.4
llama-index-cli==0.1.12
llama-index-core==0.10.36
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-ollama==0.1.2
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-ollama==0.1.3
llama-index-llms-openai==0.1.18
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.22
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.2
llamaindex-py-client==0.1.19
lxml==5.1.0
Mako==1.3.5
marisa-trie==1.1.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mutagen==1.47.0
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.30.1
opencv-python-headless==4.9.0.80
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.0
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
phonenumbers==8.13.37
pillow==10.2.0
pinecone-client==3.0.2
pluggy==1.5.0
posthog==3.3.0
preshed==3.0.9
presidio-analyzer==2.2.354
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
psygnal==0.11.1
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
py==1.11.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic-compat==0.1.2
pydantic_core==2.18.2
pydeck==0.8.1b0
PyGithub==1.59.1
Pygments==2.17.2
PyJWT==2.8.0
pylance==0.9.18
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
PyNaCl==1.5.0
pypdf==3.17.4
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyright==1.1.364
pysbd==0.3.4
PySocks==1.7.1
pytest==8.2.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
ratelimiter==1.2.0.post0
referencing==0.32.1
regex==2023.12.25
reportlab==4.2.0
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==1.3.1
retry==0.9.2
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
schema==0.7.7
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.21.0
semver==3.0.2
sentence-transformers==2.7.0
shapely==2.0.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
stack-data==0.6.3
starlette==0.32.0.post1
streamlit==1.29.0
streamlit-chat==0.0.2.2
striprtf==0.0.26
sympy==1.12
tenacity==8.2.3
thinc==8.2.3
threadpoolctl==3.5.0
tifffile==2024.5.22
tiktoken==0.5.2
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.0
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typer==0.9.0
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
validators==0.22.0
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp==3.17.0
